float32

+++ TransformersProcessor.py
@@ -13,14 +13,14 @@ class TransformersProcessor:
         
         text_inputs = self.processor(text = message, src_lang="arb", return_tensors="pt").to(self.device)
         audio_array_from_text = self.model.generate(**text_inputs, tgt_lang="arb",speaker_id=10)[0].cpu().numpy().squeeze()
-        return audio_array_from_text.astype(np.int16)
+        return audio_array_from_text.astype(np.float32)
 
     def ProcessAndWriteFile(self,message:str, file_path:str="out_from_text.wav") -> None:
         
         text_inputs = self.processor(text = message, src_lang="arb", return_tensors="pt").to(self.device)
         audio_array_from_text = self.model.generate(**text_inputs, tgt_lang="arb",speaker_id=10)[0].cpu().numpy().squeeze()
         sample_rate = self.model.config.sampling_rate
-        scipy.io.wavfile.write(file_path, rate=16000, data=audio_array_from_text.astype(np.int16))
+        scipy.io.wavfile.write(file_path, rate=16000, data=audio_array_from_text.astype(np.float32))
